backend/requirements.txt
# --- Core API ---
fastapi==0.110.1
starlette==0.37.2
uvicorn[standard]==0.25.0   # inclui watchfiles para reload em dev

# --- Banco de Dados & Migrações (SQLite síncrono) ---
SQLAlchemy==2.0.34
alembic==1.13.3

# --- Configuração & Validação ---
python-dotenv==1.1.1
pydantic==2.11.9
email-validator==2.3.0
python-multipart==0.0.20     # uploads/form-data se necessário

# --- Segurança (se for usar autenticação/jwt) ---
passlib==1.7.4               # hashing de senhas
python-jose==3.5.0           # JWT/OAuth

# --- HTTP clients (úteis p/ integrações/serviços) ---
requests==2.32.5
httpx==0.28.1

# --- Utilidades ---
anyio==4.10.0
typing_extensions==4.15.0
